refactor: move contexts to providers
