+++ src/features/NoteScreen/components/Link.tsx
@@ -3,9 +3,9 @@ import { Components } from 'react-markdown';
 import { saveAs } from 'file-saver';
 import { getAppResourceDataInUrl } from '@core/features/links';
 import { openLink } from '@electron/requests/interactions/renderer';
+import { useFilesRegistry } from '@features/Workspace/WorkspaceProvider';
 
 import { useNotesControl } from '../../MainScreen/useNotesControl';
-import { useFilesRegistry } from '../../Providers';
 
 export const Link: Exclude<Components['a'], undefined> = ({
 	node,
