+++ src/features/Profile/index.tsx
@@ -1,7 +1,6 @@
 import React, { FC } from 'react';
 import { ProfileContainer } from '@features/App/useProfiles';
 import { MainScreen } from '@features/MainScreen';
-import { dbContext } from '@features/Providers';
 import { Workspace } from '@features/Workspace';
 
 export type ProfileProps = {
@@ -12,10 +11,8 @@ export const Profile: FC<ProfileProps> = ({ profile: currentProfile }) => {
 	const activeProfileId = currentProfile.profile.id;
 
 	return (
-		<dbContext.Provider value={currentProfile.db}>
-			<Workspace profile={currentProfile}>
-				<MainScreen key={activeProfileId} />
-			</Workspace>
-		</dbContext.Provider>
+		<Workspace profile={currentProfile}>
+			<MainScreen key={activeProfileId} />
+		</Workspace>
 	);
 };
